Fixed DTO
Se arreglaron todos los servicios para que ahora devuelvan el DTO

+++ src/main/java/com/groupc/fourparks/domain/model/ParkingRate.java
@@ -11,7 +11,7 @@ import lombok.*;
 @NoArgsConstructor
 public class ParkingRate {
     private Long id;
-    private Float rate;
+    private Double rate;
     private VehicleTypeEntity vehicleTypeId;
     private ParkingEntity parkingId;
 }
